# openapi.yaml
openapi: 3.1.0
info:
  title: Convocate API
  version: 0.1.0
  description: |
    Convocate is an agentic software-engineering platform: it provisions compute
    **Nodes**, runs **Agent-containers** (Claude CLI in permission-bypass mode) on
    them, decomposes a natural-language **Specification** into a **Project Board**
    of Cards/Containers wired into execution DAGs, executes those cards against
    agents to produce code, and manages the resulting **Repositories**, **Pull
    Requests** and CI/CD through to merge.

    ## Conventions
    - All endpoints live under `/api/v1/<applet_shortname>/...`, where
      `applet_shortname` is the 2–4 letter abbreviation for the owning applet:
      `nmgr` (Node Manager), `amgr` (Agent Manager), `pb` (Project Board),
      `ide` (Code IDE), `repo` (Repo Manager), `ac` (Access Control),
      `sup` (Support Tool), plus `auth` for the session/identity service.
    - JSON request/response bodies; `application/json` unless noted.
    - Authentication via Bearer JWT (see `securitySchemes.bearerAuth`).
    - Mutating endpoints are authorized by RBAC role (see `x-required-role`);
      a principal holding the `admin` role implicitly satisfies every check.
    - List endpoints are cursor/offset paginated and return `Page<T>`.
    - Timestamps are RFC 3339 / ISO 8601 UTC strings.
servers:
  - url: https://{host}/api/v1
    variables:
      host:
        default: convocate.local
security:
  - bearerAuth: []
...
paths:
...
  /events/{applet}/{channel}:
    get:
      tags: [events]
      summary: Subscribe to real-time events for an applet channel (WebSocket upgrade)
      description: |
        Upgrades to a WebSocket. The server pushes JSON-framed events for the
        specified applet and channel. Each message is a JSON object with at
        minimum `{ type, timestamp, payload }`. The client may send
        `{ action: "ping" }` to keep the connection alive.

        Applet shortnames: `nmgr`, `amgr`, `pb`, `ide`, `repo`, `ac`, `sup`.
        Channels are applet-specific (e.g. `nmgr/status`, `amgr/status`,
        `pb/board/{boardId}`, `ide/project/{projectId}`).

        Authorization is enforced per applet: subscribing to an applet's
        events requires the applet's view role (e.g. `node-view` for `nmgr`,
        `agent-view` for `amgr`, `pb-view` for `pb`, etc.).

        This endpoint is separate from the agent shell WebSocket
        (`/amgr/agent/{agentId}/shell`), which provides bidirectional
        stdin/stdout streaming.
      x-required-role: '{applet}-view (resolved dynamically per applet)'
      parameters:
        - name: applet
          in: path
          required: true
          schema:
            type: string
            enum: [nmgr, amgr, pb, ide, repo, ac, sup]
          description: Applet shortname
        - name: channel
          in: path
          required: true
          schema:
            type: string
          description: Applet-specific channel identifier (may contain slashes)
      responses:
        '101': { description: Switching protocols (WebSocket) }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ───────────────────────────── auth ─────────────────────────────
  /auth/login:
    post:
      tags: [auth]
      summary: Authenticate with username, password and MFA token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Session established
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/oidc/github/callback:
    get:
      tags: [auth]
      summary: GitHub OIDC redirect callback; exchanges code for a session
      security: []
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
        - { name: state, in: query, required: true, schema: { type: string } }
      responses:
        '200':
          description: Session established for the federated identity
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Exchange a refresh token for a new access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties: { refreshToken: { type: string } }
      responses:
        '200':
          description: Refreshed session
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Session' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /ide/project:
    get:
      tags: [ide]
      summary: List projects
      x-required-role: ide-view
      parameters:
        - { $ref: '#/components/parameters/offset' }
        - { $ref: '#/components/parameters/limit' }
      responses:
        '200':
          description: Page of projects
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/Project' } }
    post:
      tags: [ide]
      summary: Create a project (creates a repo and a SPECIFICATION.md template)
      x-required-role: ide-update
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [name], properties: { name: { type: string } } }
      responses:
        '201': { description: Project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
  /ide/project/{projectId}/tree:
    get:
      tags: [ide]
      summary: List files and directories in a project (file explorer tree)
      x-required-role: ide-view
      parameters:
        - { name: projectId, in: path, required: true, schema: { type: string } }
        - { name: path, in: query, schema: { type: string, default: "" }, description: Directory path to list (empty for root) }
      responses:
        '200':
          description: Directory entries
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/RepoFile' } }
        '404': { $ref: '#/components/responses/NotFound' }
  /ide/project/{projectId}/file/{path}:
    parameters:
      - { name: projectId, in: path, required: true, schema: { type: string } }
      - { name: path, in: path, required: true, schema: { type: string }, description: URL-encoded repo-relative file path }
    get:
      tags: [ide]
      summary: Read a file
      x-required-role: ide-view
      responses:
        '200': { description: File, content: { application/json: { schema: { $ref: '#/components/schemas/FileContent' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [ide]
      summary: Create or overwrite a file
      x-required-role: ide-update
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [content], properties: { content: { type: string } } } } }
      responses:
        '200': { description: Saved, content: { application/json: { schema: { $ref: '#/components/schemas/FileContent' } } } }
    delete:
      tags: [ide]
      summary: Delete a file
      x-required-role: ide-update
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /ide/project/{projectId}/render-board:
    post:
      tags: [ide]
      summary: Render SPECIFICATION.md into a Project Board (LLM decomposition)
      description: |
        Sends SPECIFICATION.md to the backend LLM, which rewrites the narrative
        into a Project Board JSON of Cards and Containers, links it to the spec,
        and stores it as ProjectBoard.json in the repository.
      x-required-role: ide-update
      parameters:
        - { name: projectId, in: path, required: true, schema: { type: string } }
      responses:
        '202': { description: Board rendered, content: { application/json: { schema: { $ref: '#/components/schemas/Board' } } } }

  # ──────────────────────────── repo ──────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    offset: { name: offset, in: query, schema: { type: integer, minimum: 0, default: 0 } }
    limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 25 } }
...
  responses:
    Unauthorized:
      description: Authentication failed or missing
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Authenticated but lacking the required role
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: validation_failed }
        message: { type: string }
        details:
          type: array
          items: { type: object, properties: { field: { type: string }, issue: { type: string } } }
    Page:
      type: object
      properties:
        offset: { type: integer }
        limit: { type: integer }
        total: { type: integer }

    # ---- status ----
...
    LoginRequest:
      type: object
      required: [username, password, mfaToken]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        mfaToken: { type: string, description: TOTP token, pattern: '^[0-9]{6}$' }
    Session:
      type: object
      required: [accessToken, principal]
      properties:
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresAt: { type: string, format: date-time }
        principal: { $ref: '#/components/schemas/Principal' }
    Principal:
      type: object
      required: [id, name, roles]
      properties:
        id: { type: string }
        username: { type: string }
        name: { type: string }
        email: { type: string, format: email }
        groups: { type: array, items: { type: string } }
        roles:
          type: array
          items: { type: string }
          description: Effective roles; `admin` implies all roles.
        idp: { type: string, enum: [local, github], default: local }
        authorizedApplets:
          type: array
          items: { type: string }
          description: Applet shortnames this principal may open.

    # ---- nmgr ----
...
    CardStatus: { type: string, enum: [todo, active, done, fail, note] }
    EdgeType: { type: string, enum: [DependsOn, RelatesTo] }
    BoardSummary:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        repoId: { type: string }
        updatedAt: { type: string, format: date-time }
    Board:
      allOf:
        - $ref: '#/components/schemas/BoardSummary'
        - type: object
          properties:
            containers: { type: array, items: { $ref: '#/components/schemas/Container' } }
            cards: { type: array, items: { $ref: '#/components/schemas/Card' } }
            edges: { type: array, items: { $ref: '#/components/schemas/Edge' } }
    Geometry:
      type: object
      properties:
        x: { type: number }
        y: { type: number }
        w: { type: number }
        h: { type: number }
    Container:
      type: object
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
        agentId: { type: string, nullable: true, description: mapped agent-container }
        minimized: { type: boolean }
        geometry: { $ref: '#/components/schemas/Geometry' }
...
    Card:
      type: object
      required: [id, title, status, content, links]
      properties:
        id: { type: string }
        title: { type: string }
        status: { $ref: '#/components/schemas/CardStatus' }
        content: { type: string }
        containerId: { type: string, nullable: true }
        position: { type: object, properties: { x: { type: number }, y: { type: number } } }
        size: { type: object, properties: { w: { type: number }, h: { type: number } } }
        sourceRefs: { type: array, items: { type: string }, description: source files produced/related }
        note: { type: string, nullable: true, description: implementation note from the last run }
        links:
          type: array
          items: { $ref: '#/components/schemas/Edge' }
...
    Edge:
      type: object
      required: [type, from, to]
      properties:
        id: { type: string }
        type: { $ref: '#/components/schemas/EdgeType' }
        from: { type: string, description: source card id }
        to: { type: string, description: target card id }
...
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        repoId: { type: string }
        specificationFileId: { type: string }
        boardId: { type: string, nullable: true }
    FileContent:
      type: object
      required: [path, content]
      properties:
        path: { type: string }
        content: { type: string }
        language: { type: string }
        updatedAt: { type: string, format: date-time }

    # ---- repo ----
...
    RepoFile:
      type: object
      properties:
        name: { type: string }
        type: { type: string, enum: [file, dir] }
        size: { type: integer }
        path: { type: string }
...